[t5x] Allow custom vocab validation functions in infer() and train()
The train(), infer(), and precompile() validate that the task vocabulary matches the model vocabulary, which is generally useful to prevent config errors from vocab
mismatch. However, these make hard-coded assumptions that the model will have
one or two vocabularies and that the task will have an "inputs" and/or
"targets" vocabulary. Custom T5X models may have more than two vocabularies,
and they may not correspond to these feature names. This is the only part of
the train() and infer() functions that make such assumptions about the
underlying model.

This CL allows the user to pass in a custom vocabulary validation function in
gin, to validate vocabulary for custom models.

PiperOrigin-RevId: 452797527